Refactor model-add forms, create common superclass.
